+++ core/src/com/nanter1986/energyshooter/Enemies/LaserEnemyGreen.java
@@ -9,7 +9,7 @@ import com.nanter1986.energyshooter.Enemies.LaserOfEnemy;
  */
 
 public class LaserEnemyGreen extends LaserOfEnemy {
-    public LaserEnemyGreen(int x, int y, int spaceshipX, int screenW, int screenH) {
+    public LaserEnemyGreen(float x, float y, float spaceshipX, int screenW, int screenH) {
         super(x, y, spaceshipX, screenW, screenH);
         this.touchDamageGiven=5;
     }
